requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0
requests>=2.31.0

# LLM
google-generativeai>=0.4.0

# Browser Automation
playwright>=1.41.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# PDF Processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# Image Processing
Pillow>=10.0.0

# Visualization
matplotlib>=3.8.0

# Environment
python-dotenv>=1.0.0

# Additional utilities
beautifulsoup4>=4.12.0
lxml>=5.0.0
chardet>=5.0.0
